requirements.txt
flask
crewai
chromadb
langgraph
python-dotenv
duckduckgo-search
groq
requests
pydantic
markdown
gunicorn
edge-tts
